+++ packages/patrol_cli/lib/src/crossplatform/flutter_tool.dart
@@ -36,6 +36,7 @@ class FlutterTool {
 
   var _hotRestartActive = false;
   var _logsActive = false;
+  var _devtoolsUrl = '';
 
   /// Forwards logs and hot restarts the app when "r" is pressed.
   Future<void> attachForHotRestart({
@@ -143,6 +144,21 @@ class FlutterTool {
                 'Hot Restart for entrypoint ${basename(target)}...',
               );
               process.stdin.add('R'.codeUnits);
+            } else if (char == 'h' || char == 'H') {
+              final helpText = StringBuffer(
+                'Patrol develop key commands:\n'
+                'r Hot restart\n'
+                'h Print this help message\n'
+                'q Quit (terminate the process and application on the device)',
+              );
+
+              if (_devtoolsUrl.isNotEmpty) {
+                helpText.writeln('\nDevTools: $_devtoolsUrl');
+              } else {
+                helpText.writeln('\nDevTools: not available yet');
+              }
+
+              _logger.success(helpText.toString());
             } else if (char == 'q' || char == 'Q') {
               _logger.success('Quitting process...');
               process.kill();
@@ -171,8 +187,9 @@ class FlutterTool {
               _logger.success(
                 'Hot Restart: attached to the app\n'
                 'Patrol develop key commands:\n'
-                'r Hot restart.\n'
-                'q Quit (terminate the process and application on the device).',
+                'r Hot restart\n'
+                'h Print this help message\n'
+                'q Quit (terminate the process and application on the device)',
               );
               _hotRestartActive = true;
 
@@ -183,13 +200,13 @@ class FlutterTool {
             }
 
             if (line.startsWith('The Flutter DevTools debugger and profiler')) {
-              final devtoolsUrl = _getDevtoolsUrl(line);
+              _devtoolsUrl = _getDevtoolsUrl(line);
               _logger.success(
-                'Patrol DevTools extension is available at $devtoolsUrl',
+                'Patrol DevTools extension is available at $_devtoolsUrl',
               );
 
               if (openBrowser) {
-                unawaited(_openDevtoolsPage(devtoolsUrl));
+                unawaited(_openDevtoolsPage(_devtoolsUrl));
               }
             }