+++ packages/patrol_cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased 
 
+- Add help command `h` to `patrol develop` (#2701)
 - Add support for de, fr and pl locales for native methods that operates on strings (#2659)
 - Add support for gallery permission dialog on iOS 17 (#2659)
 - Add printing paths to the APKs after `patrol build` command. (#2685)